mcp-server/professional_tools.py
"""
MCP Tool Implementations for Meridian Professional and Manager Features
"""
import asyncio
import json
from typing import Any, Dict, List, Optional

from mcp.types import CallToolRequest, CallToolResult, TextContent


async def advanced_troubleshooting(
    self, repository: str, user_id: str, issue_description: str
) -> CallToolResult:
    """Advanced AI-powered troubleshooting for professionals"""
    try:
        # Get user context and verify role
        user_context = await self.get_user_context(user_id)
        if not user_context or user_context.role not in ['professional', 'manager']:
            return CallToolResult(
                content=[TextContent(
                    type="text", 
                    text="Access denied. This feature is only available for Professional and Manager users."
                )]
            )
        
        # Get repository analysis context
        repo_context = await self.get_repository_context(repository, user_id)
        
        # Prepare context for Gemini
        context = {
            "user_role": user_context.role,
            "repository": repository,
            "issue": issue_description,
            "repository_analysis": repo_context,
            "user_profile": {
                "github_username": user_context.github_username,
                "email": user_context.email
            }
        }
        
        prompt = f"""
You are an expert software engineering troubleshooter helping a {user_context.role} user.

Issue Description: {issue_description}
Repository: {repository}

Based on the repository analysis and issue description, provide:

1. **Root Cause Analysis**: Identify potential causes of the issue
2. **Step-by-step Debugging Guide**: Detailed debugging steps with specific commands
3. **Common Pitfalls**: What to watch out for during troubleshooting
4. **Prevention Strategies**: How to prevent similar issues in the future
5. **Tool Recommendations**: Specific tools or libraries that could help
6. **Code Examples**: If applicable, provide code snippets for fixes

Format your response as a structured troubleshooting guide that a {user_context.role} can follow.
"""
        
        ai_response = await self.call_gemini(prompt, context)
        
        return CallToolResult(
            content=[TextContent(
                type="text",
                text=f"# Advanced Troubleshooting Guide\n\n{ai_response}"
            )]
        )
        
    except Exception as e:
        return CallToolResult(
            content=[TextContent(
                type="text",
                text=f"Error in advanced troubleshooting: {str(e)}"
            )]
        )


async def performance_optimization(
    self, repository: str, user_id: str, focus_area: Optional[str] = None
) -> CallToolResult:
    """Performance optimization recommendations for professionals"""
    try:
        user_context = await self.get_user_context(user_id)
        if not user_context or user_context.role not in ['professional', 'manager']:
            return CallToolResult(
                content=[TextContent(
                    type="text",
                    text="Access denied. This feature is only available for Professional and Manager users."
                )]
            )
        
        repo_context = await self.get_repository_context(repository, user_id)
        
        context = {
            "user_role": user_context.role,
            "repository": repository,
            "focus_area": focus_area,
            "repository_analysis": repo_context
        }
        
        prompt = f"""
You are a performance optimization expert helping a {user_context.role} user optimize their codebase.

Repository: {repository}
Focus Area: {focus_area or "General performance optimization"}

Based on the repository analysis, provide:

1. **Performance Bottleneck Identification**: Key areas that likely impact performance
2. **Optimization Strategies**: Specific techniques for improvement
3. **Code-level Optimizations**: Concrete code changes with before/after examples
4. **Infrastructure Recommendations**: Deployment and scaling considerations
5. **Monitoring Setup**: Tools and metrics to track performance improvements
6. **Implementation Priority**: Which optimizations to tackle first
7. **Performance Testing**: How to measure improvements

Focus on actionable, measurable improvements that a {user_context.role} can implement.
"""
        
        ai_response = await self.call_gemini(prompt, context)
        
        return CallToolResult(
            content=[TextContent(
                type="text",
                text=f"# Performance Optimization Plan\n\n{ai_response}"
            )]
        )
        
    except Exception as e:
        return CallToolResult(
            content=[TextContent(
                type="text",
                text=f"Error in performance optimization: {str(e)}"
            )]
        )


async def best_practices_audit(
    self, repository: str, user_id: str, audit_type: str = "general"
) -> CallToolResult:
    """Best practices audit for professionals"""
    try:
        user_context = await self.get_user_context(user_id)
        if not user_context or user_context.role not in ['professional', 'manager']:
            return CallToolResult(
                content=[TextContent(
                    type="text",
                    text="Access denied. This feature is only available for Professional and Manager users."
                )]
            )
        
        repo_context = await self.get_repository_context(repository, user_id)
        
        context = {
            "user_role": user_context.role,
            "repository": repository,
            "audit_type": audit_type,
            "repository_analysis": repo_context
        }
        
        prompt = f"""
You are a senior software architect conducting a best practices audit for a {user_context.role} user.

Repository: {repository}
Audit Type: {audit_type}

Based on the repository analysis, provide a comprehensive audit covering:

1. **Code Quality Assessment**: 
   - Code organization and structure
   - Naming conventions and readability
   - Documentation quality

2. **Security Best Practices**:
   - Vulnerability assessments
   - Security configuration issues
   - Data handling practices

3. **DevOps and Deployment**:
   - CI/CD pipeline assessment
   - Deployment strategy review
   - Environment management

4. **Architecture Review**:
   - Design patterns usage
   - Scalability considerations
   - Maintainability factors

5. **Testing Strategy**:
   - Test coverage analysis
   - Testing approach evaluation
   - Quality assurance practices

6. **Action Plan**:
   - Priority-ranked improvements
   - Implementation timeline
   - Resource requirements

Provide specific, actionable recommendations with examples where applicable.
"""
        
        ai_response = await self.call_gemini(prompt, context)
        
        return CallToolResult(
            content=[TextContent(
                type="text",
                text=f"# Best Practices Audit Report\n\n{ai_response}"
            )]
        )
        
    except Exception as e:
        return CallToolResult(
            content=[TextContent(
                type="text",
                text=f"Error in best practices audit: {str(e)}"
            )]
        )


async def advanced_learning_suggestions(
    self, repository: str, user_id: str, skill_focus: Optional[str] = None
) -> CallToolResult:
    """Advanced learning recommendations for professionals"""
    try:
        user_context = await self.get_user_context(user_id)
        if not user_context or user_context.role not in ['professional', 'manager']:
            return CallToolResult(
                content=[TextContent(
                    type="text",
                    text="Access denied. This feature is only available for Professional and Manager users."
                )]
            )
        
        repo_context = await self.get_repository_context(repository, user_id)
        
        context = {
            "user_role": user_context.role,
            "repository": repository,
            "skill_focus": skill_focus,
            "repository_analysis": repo_context
        }
        
        prompt = f"""
You are an expert learning and development advisor for a {user_context.role} software developer.

Repository: {repository}
Skill Focus: {skill_focus or "General professional development"}

Based on the repository analysis and current skill level, provide:

1. **Skill Gap Analysis**: 
   - Current skill level assessment
   - Areas for improvement identified from codebase
   - Industry trends alignment

2. **Learning Path Recommendations**:
   - Advanced topics to master
   - Specific technologies to learn
   - Certification opportunities

3. **Practical Projects**:
   - Hands-on projects to build skills
   - Open source contributions to consider
   - Portfolio enhancement ideas

4. **Resources and Materials**:
   - Books, courses, and documentation
   - Conferences and workshops
   - Community and networking opportunities

5. **Implementation Timeline**:
   - Short-term goals (1-3 months)
   - Medium-term objectives (3-6 months)
   - Long-term career development (6+ months)

6. **Application to Current Project**:
   - How to apply new skills to current repository
   - Refactoring opportunities for practice
   - Architecture improvements to explore

Focus on advanced, professional-level content that will accelerate career growth.
"""
        
        ai_response = await self.call_gemini(prompt, context)
        
        return CallToolResult(
            content=[TextContent(
                type="text",
                text=f"# Advanced Learning Development Plan\n\n{ai_response}"
            )]
        )
        
    except Exception as e:
        return CallToolResult(
            content=[TextContent(
                type="text",
                text=f"Error in learning suggestions: {str(e)}"
            )]
        )


# These functions are imported and used by the MCP server
