mcp-server/requirements.txt
# Core AI and API dependencies
google-generativeai>=0.8.3
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Server dependencies  
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlite3-utils>=3.34

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
